+++ Futility/UIViewController+FKSplashScreenAnimation.m
@@ -83,7 +83,7 @@ static UIWindow *splashWindow;
 #define CHANGE_X(_view, _x) (_view).frame = CGRectMake((_x), (_view).frame.origin.y, (_view).frame.size.width, (_view).frame.size.height)
 #define ANIMATION_IS_USED(_animationsBitfield, _animationBit) ((_animationsBitfield) & (_animationBit))
 
-static BOOL animationIncludesSliding(unsigned int animations)
+static BOOL animationIncludesSliding(NSUInteger animations)
 {
     return (ANIMATION_IS_USED(animations, FKSplashScreenAnimation_SlideUp)
             || ANIMATION_IS_USED(animations, FKSplashScreenAnimation_SlideDown)
@@ -92,7 +92,7 @@ static BOOL animationIncludesSliding(unsigned int animations)
             );
 }
 
-- (void) fk_prepareAnimations:(unsigned int)animations forView:(UIView *)splashView
+- (void) fk_prepareAnimations:(NSUInteger)animations forView:(UIView *)splashView
 {
     if (animationIncludesSliding(animations))
     {
@@ -110,7 +110,7 @@ static BOOL animationIncludesSliding(unsigned int animations)
     }
 }
 
-- (void) fk_performAnimations:(unsigned int)animations forView:(UIView *)splashView
+- (void) fk_performAnimations:(NSUInteger)animations forView:(UIView *)splashView
 {
     if (ANIMATION_IS_USED(animations, FKSplashScreenAnimation_SlideUp))
         CHANGE_Y(splashView, -(splashView.frame.size.height + splashView.layer.shadowRadius));
@@ -137,7 +137,7 @@ static BOOL animationIncludesSliding(unsigned int animations)
     }
 }
 
-- (void) fk_completeAnimations:(unsigned int)animations forView:(UIView *)splashView
+- (void) fk_completeAnimations:(NSUInteger)animations forView:(UIView *)splashView
 {
     self.view.transform = CGAffineTransformIdentity;
 }
@@ -149,7 +149,7 @@ static BOOL splashScreenHasBeenAnimated = NO;
 }
 
 - (UIImageView *) fk_animateSplashScreenRemovalWithDuration:(NSTimeInterval)duration
-                                                 animations:(unsigned int)animations
+                                                 animations:(NSUInteger)animations
                                                  completion:(void(^)(BOOL finished))completion
 {
     if (animations == FKSplashScreenAnimation_None)
